slow the animation down without touching physics
Noodle asked me to slow down the animation without breaking the physics,
so I did what I could.

I faked the slower animation by adding a 'fake' fishing rod for the
duration of the 'fish' animation, and animating that at a slower speed.
The original fishing rod is untouched, so physics should be unaffected.
As the objects are still relatively fast-moving, the rope end detaching
from the fishing rod momentarily is (hopefully) not too noticeable
in-game.

Wasn't able to properly test this as
1. it is 02:30 as I write this
2. the game lags quite badly on my machine due to the log spam and the
   water shader doing something funky

Hopefully this is an improvement.